backend/requirements.txt
django>=4.2
djangorestframework>=3.14
channels>=4.0
docker>=7.0
channels-redis>=4.1
django-cors-headers>=4.0.0
